move enums to their own file
