+++ src/enums/evolution-condition-type.ts
@@ -1,0 +1,25 @@
+export enum EvolutionConditionType {
+  NONE,
+  DAY,
+  NIGHT,
+  FRIENDSHIP,
+  FRIENDSHIP_DAY,
+  FRIENDSHIP_NIGHT,
+  MOVE,
+  MOVE_DAY,
+  MOVE_NIGHT,
+  MOVE_TYPE,
+  FRIENDSHIP_MOVE_TYPE,
+  PARTY_MEMBER_TYPE,
+  CAUGHT_SPECIES,
+  WEATHER,
+  TREASURE,
+  NATURE,
+  MALE,
+  FEMALE,
+  BIOME,
+  TYROGUE,
+  SHEDINJA,
+  DUNSPARCE,
+  TANDEMAUS
+}
