+++ src/enums/species-wild-evolution-delay.ts
@@ -1,0 +1,8 @@
+export enum SpeciesWildEvolutionDelay {
+  NONE,
+  SHORT,
+  MEDIUM,
+  LONG,
+  VERY_LONG,
+  NEVER
+}
